Merge branch 'force_reinstall_docutils' into 'master'
Don't be cautious to preserve docutils build directory pip.

Pip is cautious on removing build directories:

```
pip can't proceed with requirements 'docutils==0.11' due to a pre-existing build directory (/var/cache/omnibus/src/python-docutils/docutils). This is likely due to a previous installation that failed. pip is being responsible and not assuming it can delete this. Please delete it and try again.
```

We don't really care about the cache directories so allow it to remove.

See merge request !751