Merge branch 'mail_room_timeout' into 'master'
Change the default imap idle timeout of 29 minutes to 60 seconds

This will fix the mail_room hanging issue. Since most IMAP servers are not keeping connections open for 29 minutes. And even when a mailserver support that a firewall in between can stop the connection after X amount of minutes anyway.

See merge request !842