Run the mattermost version check after the config and log directory have been created
The mattermost recipe was checking the mattermost version before the config and directories were put in place.

In addition our migration error logging for mattermost now logs as the mattermost user. Previously, on error the log may not have been created before this error message, and logging as root would cause a permission problem on the file for future runs.