1345 Add option to disable migrations completely
Add option to ``gitlab.rb`` to enable/disable the database migration steps as part of install/upgrade/reconfigure.

Adds: ``gitlab_rails['auto_migrate'] (=true)`` to the system, and an ``only_if`` conditional on this variable to the ``database_migration.rb`` book. If this value is set to false, it will skip the migration steps even if there is no previous record of a migration having successfully having happened.