Revise packagecloud_upload (verbosity, readability)
For verbosity, move the embedded `find` command into a `for` loop, allowing us to specifially outout what files are being pushed to packagecloud, as well as making it more directly legible in the future.

To the readability aspect, I have also renamed the variables in the for
loops from `x` to `distro`, and added `package` vs `y`.