Add the same security patch for git that we did for stable
We need to do this again because this we are now
shipping git 2.8 instead of git 2.7, and the previous
patch needed to be updated to work.

Instead of updating the previus patch, I have reported the
original patch from http://public-inbox.org/git/20161201090336.xjbb47bublfcpglo@sigill.intra.peff.net/

And seen that is is essentially the same as the previous.