Pass both sets of namespaces and tags for image retagging
Since we are now uploading docker images to a temporary repo first and then
pull from there to push to dockerhub, retagging is no longer simply changing
the tag. Retagging now involves changing the namespace also. Earlier, it was
w:x to w:y, but now, it is w:x to m:n