Update unicorn worker to use 400MB as the variable in the per process calculation
Unicorn workers currently are set to be reaped between 400 and 650MB.
Which were bumped up from 350MB a while back, but our caclulation wasn't
updated

Update the rounding fix in the numerator of the caclulation, which
hasn't been updated in a long time. And improve the comment so it isn't
missed in the future.

Reserved space has been bumped from 1GB to 1.5GB. This is 512MB for the
OS (the ubuntu minimum when not running Unity) and 1GB for the rest of
GitLab.

For determining what the rest of gitlab required, I ran `docker run
gitlab/gitlab-ee:nightly` then execed in and turned off unicorn using
`gitlab-ctl stop unicorn`

Then I ran `docker stats` to see the memory usage of the processes
running in the container. These were sitting just above 900MB.