Add `quote` function to templatesymlink managed templates
In order to correctly handle the emitting of strings correctly escaped
for use in a YAML file, we add a `quote` helper function, and remove the
older `single_quote` helper.

* Adds QuoteHelper with new `quote` function
* Adds spec examples `quote` helper function
* Converts existing uses of `single_quote` to `quote`
* Removes SingleQuoteHelper

This may cause a slight bump if external code is using SingleQuoteHelper,
though these are all internal cookbooks to manage omnibus-gitlab.