Disable prepared statements by default
For GitLab.com the use of prepared statements did not yield any
benefits. Furthermore we have had reports of users experiencing memory
leaks due to what appeared to be the use of prepared statements. Some
details of this can be found in
https://gitlab.com/gitlab-org/gitlab-ce/issues/20723. Prepared
statements also don't work when using pgbouncer in transaction pooling
mode (the mode we recommend you to use).

Because of all this we have decided to disable prepared statements by
default, preventing users from running into nasty issues when using
pgbouncer or similar software.