Ask gem where it put `gitlab-markup`
In order to reduce the fragile nature of the patching, ask `gem` what
the contents of `gitlab-markup` are, and then look for
lib/github/markups.rb within that file list.  This will protect us from
minor version bumps to gitlab-markup, as they should keep the same file
structure between releases.