Remove rsync instruction from promote-to-primary
The promote-to-primary-node command is used for disaster recovery
using geo to recover quickly. It is possible the primary may never be
recovered, but at the very least it is almost certain that the primary
is not currently accessible.

This command should provide feedback that relates to the process of
making the GitLab instance accessible again. Additional data recovery
tasks, particularly those requiring the old primary to be brought back
online, are secondary and should be documented later in the process.