Update template `postgresql['custom_hba_entries']`
In 6dd76e32a8351e50f4ecbbb1d82aad2cf24fc896 the
postgresql['custom_hba_entries'] setting was changed from being an Array
of Hashes to a Hash of Arrays of Hashes

In 858706629e593bd3fb3393e23f33e6a079416474 the example in the template
was changed to demonstrate a Hash of Hashes.

Here we take one more pass to make the template match the current
implementation.