Geo: Disable SSL compression by default in generating recovery.conf
We saw that on GitLab.com, turning off SSL compression while keeping SSL on
virtually eliminated replication lag while the site was operating. This
change won't affect a current Geo installation since this requires a change
in recovery.conf and a PostgreSQL restart on the secondary, but it will
take effect for new installations.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/issues/3495