Use letsencrypt for external registry certificate
Currently when configuring an install we expect distinct certificates
for the main address and the registry.

Here we detect the situation where a certficate file does not exist for
the the registry.  When this is the case we request an alt_name be added
to the main certifcate managed by letsencrypt, and and configure
nginx_registry to use the main certificate.