Add another tip to Geo replication troubleshooting
It seems people regularly run into the problem where they forget a
step in the setup of Geo, so they get the following error:

> ERROR:  replication slots can only be used if wal_level >= replica

This error seems to happen when on the primary to role was not
configured correctly. The admin should have added
`roles ['geo_primary_role']` or `geo_primary_role['enable'] = true` to
`gitlab.rb` on the primary node.

Since this is a _common_ error, add this as a tip to the
troubleshooting section of the error message.