Docker: attempt clean stale pids on restart
Issue 4235 points to a problem with being unable to properly restart
the omnibus containers. One of the cases is in the even of server
outage (or unclean stop).

This change ads the functionality of cleaning up PIDs & socket files
prior to the start of runsvdir.

This _does not_ address `...: runsv not running` complaints directly.