Add Rugged search path to Gitaly config
By default, Rugged searches inside `/etc/gitconfig` instead of
`/opt/gitlab/embedded/etc/gitconfig` for system-wide options unless the
`Rugged::Settings['search_path_system']` is set.

This means that important options such as `core.fsyncObjectFiles` are
not propagated to gitaly-ruby, which can result in data loss after
servers are rebooted. We worked around this problem by propagating the
`core` options in the user's `.gitconfig` file in
https://gitlab.com/gitlab-org/omnibus-gitlab/merge_requests/3648, but we
shouldn't need to do this.

The command-line `git` doesn't have this issue because it uses the
`prefix` compile-time option, but this isn't available in Rugged.