Raise unicorn memory limits
The base memory usage of a Unicorn worker is about 500-600 MB now. The
previous memory limits (min: 400, max: 600) were too low because Unicorn
Worker Killer picks a random number between 400 and 600 MB and will kill
the process if it happens to fall in that number. As a result, there was
a high probability that a Unicorn worker would be killed after **every**
request.

To fix this, we raise the limits to match GitLab.com: 1 GB and 1.28 GB.

We also tweak the formula for calculating number of Unicorn workers. This
increase of baseline RAM lowers the number of workers.

Closes https://gitlab.com/gitlab-org/omnibus-gitlab/issues/4279