Allow enabling experimental sidekiq-cluster
We're moving to running Sidekiq cluster as a default in the
future. This allows administrators to opt-in to the sidekiq cluster
behaviour already.

This behaviour will copy the configuration of sidekiq to
sidekiq-cluster. So it will respect concurrency and logging settings.

When enabling `cluster` for Sidekiq, this will spin up a single
Sidekiq-cluster and Sidekiq process. Sidekiq-cluster will be
configured with the queue groups set to [`*`]. Which will spin up a
single Sidekiq process picking up all queues.