Add get-postgresql-primary command to gitlab-ctl
`get-postgresql-primary` command prints out the connection information
(IP address and port) of primary PostgreSQL node. This is useful to
identify the primary node in a clustered deployment, for example with
repmgr or Patroni.

This command requires a Consul agent on the same node that it is
running. It uses Consul DNS interface to lookup the SRV record for
PostgreSQL service.