Use gcc v6.3 for CentOS 6 to build Rails C extensions
This will help unblock the upgrade to Ruby 2.7 since gRPC and
google-protobuf need a more recent compiler.

As explained in https://stackoverflow.com/a/15602073/1992201:

The whole point of devtoolset is that it allows you to use a newer GCC
and newer libstdc++ but without creating any run-time dependencies on
the newer libstdc++ library. The compiled executables only need the
system version of libstdc++.so that is always present on RHEL/CentOS,
even systems without devtoolset installed. (What devtoolset does is
package all the new libstdc++ features in a static library, called
libstdc++_nonshared.a so that all the pieces not present in the system
libstdc++.so get statically linked in, and everything else will come
from the system libstdc++.so).