Only send one Referrer-Policy header
We send Referrer-Policy from nginx based on user configuration. However,
Rails also sends this header by default, with a value of
`strict-origin-when-cross-origin`.

We want to keep the default from Rails (for non-Omnibus or non-nginx
installs), but only send a single value for this header in Omnibus nginx
situations. Sending two values is not technically a problem, as long as
the configured value comes last[1], but it is potentially confusing and
order-dependent, so this makes it clearer with no change in behaviour.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referrer-Policy#Specifying_a_fallback_policy