Set net.core.somaxconn kernel parameter for Puma
Set net.core.somaxconn kernel parameter to a value equal or higher then
the default Puma backlog_socket. The default kernel value is 128, which
is lower then the default Puam backlog.

This was done for Unicorn in
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/405, but
this was not done for Puma. Installations after GitLab 13.0+ might see
lower throughput if this setting is not raised.