configure sshd and git account so that pam is not used any longer
The reason is that some container systems don't run containers with
CAP_AUDIT_WRITE which is needed for pam. As the comment in sshd_config
has pointed out pam was used as a Fix because it handles shadow entries
'!' and '*' differently from ssh. This commit refixes that fix and
creates the git account with '*' (no password login) and disables pam in
sshd_config

Fixes #5891