Fix deployer feature gate only for auto-deploy
With
https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/4970 we
introduced a feature gate on the gitlab_com:deployer task to make it a
no-op if OMNIBUS_TRIGGERS_DEPLOYER is set to false.

This has the side effect of preventing deployments to other
environments as well like preprod or releases.

This commit changes the feature gate to check
AUTO_DEPLOY_OMNIBUS_TRIGGERS_DEPLOYER and only perform a no-op in case
the package is an auto-deploy build.