Fix deprecation logic which could auto-enable services
Due to the way Gitlab::ConfigMash works, when auto-vivifaction is
enabled, `[]` method will create an empty Hash if the specified key is
not present. This means in scenarios where unicorn['enable'] is not at
all set in gitlab.rb, when we try to fetch unicron['enable'] to check
for deprecations, because the key `enable` is not present, it will cause
unicorn['enable'] to be an empty Hash.  Because we check for `nil`-ness
of service['enable'] to include enable/disable recipes and because an
empty hash is not `nil`, we will end up enabling the unicorn service,
and boom!

Signed-off-by: Balasankar "Balu" C <balasankar@gitlab.com>
Changelog: fixed